Restore password database authentication
