+++ pyproject.toml
@@ -58,7 +58,6 @@ server = [
     "uuid-utils>=0.10",
     "cryptography>=49.0.0",
     "packaging>=24.0",
-    "boto3>=1.42",
 ]
 worker = [
     "pydantic-settings>=2.14",
